integração/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
kerykeion
svgwrite

